Adds initial setup for shared base linting package
This adds:
- a `package.json` with basic config, a script to lint the package
itself, and dependencies needed for consuming apps to use basic linting.
- a local `.eslintrc` for linting the package itself
- a `.prettierrc` for editors that use prettier. Since we also want to
export prettier rules as a shared part of the package, this includes the
`eslint-plugin-prettier` package which allows us to define prettier rules
alongside eslint rules in `index.js`. This means consuming apps will be
able to share the same prettier rule validation.