Changes the ordering of 'extends' to mimic original project
We want to maintain parity with the ordering of the original project's
linting files. This moves around the ordering of the `extends` to mimic
the original eslint files, which should then apply lint rules in the
same way. It also moves individual `extends` out of `/rules` files so we
have direct control over which rules get applied first.