Update Rule in eslint-config-ezcater-react
We were seeing a conflict between Prettier and the jsx-wrap-multilines
rule when passing components as props to other components.
React/jsx-wrap-multilines wanted to wrap multiline components in
parentheses, but Prettier didn't feel that they were necessary.

To fix this, we're going to turn the jsx-wrap-multilines rule off. We
discovered that Pretter enforces the exact style that we want, so this
will be a safe change.

The version for the react linting rules had to be bumped for this, and
since the typescript package depends on the react package, that one has
been bumped as well.