fix: copy prev rules from eslint-plugin-prettier
What did we change?

Disable rules `arrow-body-style` and `prefer-arrow-callback` to copy
previous behavior from `eslint-plugin-prettier`.

Why are we doing this?

With PR #6 `eslint-plugin-prettier` was dropped as a dependency.
`eslint-plugin-prettier` turns off the rules: `arrow-body-style`,
`prefer-arrow-callback`.

`arrow-body-style` is a rule that will cause a lot of conflicts in
existing builds. `prefer-arrow-callback` is less likely to cause issues,
but it's being turned off just because that was existing behavior when
we had eslint-plugin-prettier.