build!: move plugins to peer deps for base config
What did we change?
Move `eslint-plugin-import` from dependencies to requiring as a peer
dependency (`eslint-plugin-import` is a peer dependency of
`eslint-config-airbnb-base`). Clients are now expected to add the
dependency themselves as a dev dependency.

Why are we doing this?
This is following eslint's best practices for shareable configs of
adding plugins as a peer dependencies.